+++ NEWS.md
@@ -1,3 +1,7 @@
 # geR2 0.1.0
 
-* Added a `NEWS.md` file to track changes to the package.
+* Added a `NEWS.md` file to track changes to the package  
+* Added `ger_proj()`, `ger_code()`, `ger_data()`, `ger_docs()`, and `ger_setup()`   
+* Added `data-raw/wu_df.R`, `data-raw/wu_dt.R`, and `data-raw/wu_tbl.R`  
+* Added `inst/extdata/wu_data.csv`  
+* Added `inst/rmarkdown/templates/ger2-README/README.Rmd` for `README.Rmd` template